requirements.txt
fastapi==0.95.0
gunicorn==20.1.0
Flask==2.2.2
numpy==1.23.2
pillow==9.5.0
Requests==2.31.0
tensorflow==2.12.0
uvicorn==0.21.1
python-multipart==0.0.5